+++ libraries/AP_ExternalAHRS/AP_ExternalAHRS_config.h
@@ -10,8 +10,12 @@
 #define AP_EXTERNAL_AHRS_BACKEND_DEFAULT_ENABLED HAL_EXTERNAL_AHRS_ENABLED
 #endif
 
-#ifndef AP_EXTERNAL_AHRS_MICROSTRAIN_ENABLED
-#define AP_EXTERNAL_AHRS_MICROSTRAIN_ENABLED AP_EXTERNAL_AHRS_BACKEND_DEFAULT_ENABLED
+#ifndef AP_EXTERNAL_AHRS_MICROSTRAIN5_ENABLED
+#define AP_EXTERNAL_AHRS_MICROSTRAIN5_ENABLED AP_EXTERNAL_AHRS_BACKEND_DEFAULT_ENABLED
+#endif
+
+#ifndef AP_MICROSTRAIN_ENABLED
+#define AP_MICROSTRAIN_ENABLED AP_EXTERNAL_AHRS_MICROSTRAIN5_ENABLED
 #endif
 
 #ifndef AP_EXTERNAL_AHRS_VECTORNAV_ENABLED
